feat: apply selectedRegion
