+++ packages/frontend/features/results/sections/TitleSection.tsx
@@ -6,6 +6,8 @@ import React, { useEffect, useMemo } from 'react';
 
 import RankArrowButton from '@/components/atoms/buttons/RankArrowButton';
 import JewelRadarChart from '@/components/atoms/charts/RadarChart';
+import KlaciCodeCircles from '@/components/atoms/circle/KlaciCodeCircles';
+import { generateDistrictPath } from '@/utils/districtPath';
 
 interface DistrictData {
   id: string;
@@ -17,64 +19,16 @@ interface TitleSectionProps {
   districtData?: DistrictData | null;
 }
 
-// KLACI Code 원형 컴포넌트
-const KlaciCodeCircles: React.FC<{ klaciCode: string }> = ({ klaciCode }) => {
-  // KLACI 코드 색상 매핑 (대문자 기준)
-  const codeColorMapping: Record<string, string> = {
-    G: '#FF3737',
-    T: '#FFA600',
-    V: '#874FFF',
-    R: '#24CB71',
-  };
-
-  // klaci_code 문자열을 개별 문자로 분리
-  const klaciCodes = klaciCode.split('').map((code) => ({
-    code, // 원본 코드 그대로 표시 (소문자면 소문자로)
-    color: codeColorMapping[code.toUpperCase()] || '#666666', // 색상 매칭은 대문자로 변환
-  }));
-
-  return (
-    <div
-      style={{
-        display: 'flex',
-        gap: '8px',
-        marginBottom: '24px',
-      }}
-    >
-      {klaciCodes.map((item, index) => (
-        <div
-          key={index}
-          style={{
-            width: '30px',
-            height: '30px',
-            borderRadius: '50%',
-            backgroundColor: item.color,
-            color: 'white',
-            border: `2px solid ${item.color}`,
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            fontSize: '1rem',
-            fontWeight: 600,
-          }}
-        >
-          {item.code}
-        </div>
-      ))}
-    </div>
-  );
-};
-
 const TitleSection: React.FC<TitleSectionProps> = ({ districtData }) => {
   const router = useRouter();
 
   // Zustand store에서 선택된 지역 정보 가져오기
-  const { selectedDistrict } = useDistrict();
+  const { selectedRegion } = useDistrict();
   const getProvinceById = useGetProvinceById();
 
   // 차트 데이터를 동적으로 생성하는 함수
-  const generateChartData = (district: any): number[] => {
-    if (!district) {
+  const generateChartData = (region: any): number[] => {
+    if (!region) {
       // 기본값 반환
       return [50, 50, 50, 50, 50, 50, 50, 50];
     }
@@ -84,7 +38,7 @@ const TitleSection: React.FC<TitleSectionProps> = ({ districtData }) => {
       economy_score = 50,
       living_score = 50,
       safety_score = 50,
-    } = district;
+    } = region;
 
     // RadarChart의 categories 순서에 맞춰 반환:
     // ['생활역동형', '안전복원형', '인구정착형', '경제정속형', '생활정주형', '안전정진형', '인구성장형', '경제혁신형']
@@ -100,205 +54,17 @@ const TitleSection: React.FC<TitleSectionProps> = ({ districtData }) => {
     ];
   };
 
-  // selectedDistrict가 변경될 때마다 차트 데이터 재계산
+  // selectedRegion가 변경될 때마다 차트 데이터 재계산
   const chartData = useMemo(() => {
-    return generateChartData(selectedDistrict);
-  }, [selectedDistrict]);
+    return generateChartData(selectedRegion);
+  }, [selectedRegion]);
 
-  // URL path 생성 함수
-  const generateDistrictPath = (district: any): string => {
-    if (!district?.name || !district?.province_id) {
-      return 'jeonbuk-jeonju'; // 기본값
-    }
-
-    const province = getProvinceById(district.province_id);
-    if (!province?.name) {
-      return 'jeonbuk-jeonju'; // 기본값
-    }
-
-    // 한글-영문 매핑
-    const provinceMapping: Record<string, string> = {
-      강원특별자치도: 'gangwon',
-      경기도: 'gyeonggi',
-      경상남도: 'gyeongnam',
-      경상북도: 'gyeongbuk',
-      광주광역시: 'gwangju',
-      대구광역시: 'daegu',
-      대전광역시: 'daejeon',
-      부산광역시: 'busan',
-      서울특별시: 'seoul',
-      세종특별자치시: 'sejong',
-      울산광역시: 'ulsan',
-      인천광역시: 'incheon',
-      전라남도: 'jeonnam',
-      전라북도: 'jeonbuk',
-      제주특별자치도: 'jeju',
-      충청남도: 'chungnam',
-      충청북도: 'chungbuk',
-    };
-
-    // 지역별 매핑 (province와 district를 함께 고려)
-    const getDistrictMapping = (
-      provinceName: string,
-      districtName: string,
-    ): string => {
-      // 서울특별시
-      if (provinceName === '서울특별시') {
-        const seoulMapping: Record<string, string> = {
-          종로구: 'jongno',
-          중구: 'jung-seoul',
-          용산구: 'yongsan',
-          성동구: 'seongdong',
-          광진구: 'gwangjin',
-          동대문구: 'dongdaemun',
-          중랑구: 'jungnang',
-          성북구: 'seongbuk',
-          강북구: 'gangbuk',
-          도봉구: 'dobong',
-          노원구: 'nowon',
-          은평구: 'eunpyeong',
-          서대문구: 'seodaemun',
-          마포구: 'mapo',
-          양천구: 'yangcheon',
-          강서구: 'gangseo',
-          구로구: 'guro',
-          금천구: 'geumcheon',
-          영등포구: 'yeongdeungpo',
-          동작구: 'dongjak',
-          관악구: 'gwanak',
-          서초구: 'seocho',
-          강남구: 'gangnam',
-          송파구: 'songpa',
-          강동구: 'gangdong',
-        };
-        return (
-          seoulMapping[districtName] ||
-          districtName
-            .toLowerCase()
-            .replace(/시|군|구/g, '')
-            .replace(/\s+/g, '')
-        );
-      }
-
-      // 인천광역시
-      if (provinceName === '인천광역시') {
-        const incheonMapping: Record<string, string> = {
-          중구: 'jung-incheon',
-          동구: 'dong-incheon',
-          미추홀구: 'michuhol',
-          연수구: 'yeonsu',
-          남동구: 'namdong',
-          부평구: 'bupyeong',
-          계양구: 'gyeyang',
-          서구: 'seo-incheon',
-          강화군: 'ganghwa',
-          옹진군: 'ongjin',
-        };
-        return (
-          incheonMapping[districtName] ||
-          districtName
-            .toLowerCase()
-            .replace(/시|군|구/g, '')
-            .replace(/\s+/g, '')
-        );
-      }
-
-      // 경기도
-      if (provinceName === '경기도') {
-        const gyeonggiMapping: Record<string, string> = {
-          수원시: 'suwon',
-          성남시: 'seongnam',
-          의정부시: 'uijeongbu',
-          안양시: 'anyang',
-          부천시: 'bucheon',
-          광명시: 'gwangmyeong',
-          평택시: 'pyeongtaek',
-          동두천시: 'dongducheon',
-          안산시: 'ansan',
-          고양시: 'goyang',
-          과천시: 'gwacheon',
-          구리시: 'guri',
-          남양주시: 'namyangju',
-          오산시: 'osan',
-          시흥시: 'siheung',
-          군포시: 'gunpo',
-          의왕시: 'uiwang',
-          하남시: 'hanam',
-          용인시: 'yongin',
-          파주시: 'paju',
-          이천시: 'icheon',
-          안성시: 'anseong',
-          김포시: 'gimpo',
-          화성시: 'hwaseong',
-          광주시: 'gwangju-gyeonggi',
-          여주시: 'yeoju',
-          양평군: 'yangpyeong',
-          고양군: 'goyang-gun',
-          연천군: 'yeoncheon',
-          가평군: 'gapyeong',
-          포천군: 'pocheon',
-        };
-        return (
-          gyeonggiMapping[districtName] ||
-          districtName
-            .toLowerCase()
-            .replace(/시|군|구/g, '')
-            .replace(/\s+/g, '')
-        );
-      }
-
-      // 전라북도
-      if (provinceName === '전라북도') {
-        const jeonbukMapping: Record<string, string> = {
-          전주시: 'jeonju',
-          군산시: 'gunsan',
-          익산시: 'iksan',
-          정읍시: 'jeongeup',
-          남원시: 'namwon',
-          김제시: 'gimje',
-          완주군: 'wanju',
-          진안군: 'jinan',
-          무주군: 'muju',
-          장수군: 'jangsu',
-          임실군: 'imsil',
-          순창군: 'sunchang',
-          고창군: 'gochang',
-          부안군: 'buan',
-        };
-        return (
-          jeonbukMapping[districtName] ||
-          districtName
-            .toLowerCase()
-            .replace(/시|군|구/g, '')
-            .replace(/\s+/g, '')
-        );
-      }
-
-      // 기본 처리
-      return districtName
-        .toLowerCase()
-        .replace(/시|군|구/g, '')
-        .replace(/\s+/g, '');
-    };
-
-    // province name을 영문으로 변환
-    const provinceName =
-      provinceMapping[province.name] ||
-      province.name
-        .toLowerCase()
-        .replace(/특별자치도|광역시|특별시/g, '')
-        .replace(/\s+/g, '');
-
-    // district name을 영문으로 변환 (province 정보를 함께 고려)
-    const districtName = getDistrictMapping(province.name, district.name);
-
-    return `${provinceName}-${districtName}`;
-  };
-
-  // selectedDistrict 변경 시 URL 업데이트
+  // selectedRegion 변경 시 URL 업데이트
   useEffect(() => {
-    if (selectedDistrict) {
-      const newPath = generateDistrictPath(selectedDistrict);
+    if (selectedRegion) {
+      const newPath = generateDistrictPath(selectedRegion, {
+        getProvinceById,
+      });
       console.log('Updating URL to:', newPath);
 
       // 현재 URL과 다를 때만 업데이트 (무한 루프 방지)
@@ -307,31 +73,28 @@ const TitleSection: React.FC<TitleSectionProps> = ({ districtData }) => {
         router.replace(`/results/${newPath}`, { scroll: false });
       }
     }
-  }, [selectedDistrict, router, getProvinceById]);
+  }, [selectedRegion, router, getProvinceById]);
 
   // 안전한 지역명 생성 함수
   const getDistrictName = (): string => {
-    console.log('Selected district:', selectedDistrict);
-
-    // selectedDistrict가 유효한 객체이고 name 속성이 있는 경우
-    if (selectedDistrict?.name && selectedDistrict?.province_id) {
-      // province_id를 사용해서 province 정보 가져오기
-      const province = getProvinceById(selectedDistrict.province_id);
-      console.log('Found province:', province);
-
-      if (province?.name) {
-        return `${province.name} ${selectedDistrict.name}`;
-      }
+    if (selectedRegion?.province?.name && selectedRegion?.name) {
+      return `${selectedRegion.province.name} ${selectedRegion.name}`;
     }
-
-    // selectedDistrict가 없거나 유효하지 않은 경우 기본값 반환
-    return '전라북도 전주시';
+    // selectedRegion이 없거나 유효하지 않은 경우 기본값 반환
+    return 'NOT SELECTED';
   };
 
   // 기본값 설정
-  const rank = districtData?.rank || 3;
+  // const rank = districtData?.rank || selectedRegion?.total_rank || 3;
+  const rank = useMemo(() => {
+    return selectedRegion?.total_rank || 3;
+  }, [selectedRegion]);
   const rankText = `종합순위 ${rank}위`;
   const districtName = getDistrictName();
+
+  // KLACI 코드와 닉네임 가져오기
+  const klaciCode = selectedRegion?.klaci?.code || 'KLAC';
+  const klaciNickname = selectedRegion?.klaci?.nickname || '안전복지형';
 
   // 다음/이전 지자체로 이동하는 함수
   const handleNavigate = (direction: 'prev' | 'next') => {
@@ -430,10 +193,10 @@ const TitleSection: React.FC<TitleSectionProps> = ({ districtData }) => {
         {districtName}
       </div>
 
-      {/* KLACI Code 원형 컴포넌트 - selectedDistrict의 klaci_code 사용 */}
-      <KlaciCodeCircles klaciCode={selectedDistrict?.klaci_code || 'KLAC'} />
+      {/* KLACI Code 원형 컴포넌트 - selectedRegion의 klaci 정보 우선 사용 */}
+      <KlaciCodeCircles klaciCode={klaciCode} />
 
-      {/* 유형 텍스트 */}
+      {/* 유형 텍스트 - selectedRegion의 klaci 닉네임 사용 */}
       <div
         style={{
           fontSize: '2.2rem',
@@ -442,7 +205,7 @@ const TitleSection: React.FC<TitleSectionProps> = ({ districtData }) => {
           marginBottom: '12px',
         }}
       >
-        안전복지형
+        {klaciNickname}
       </div>
 
       {/* 유형 설명 */}
