+++ packages/frontend/package.json
@@ -20,6 +20,8 @@
   "dependencies": {
     "@nivo/core": "^0.99.0",
     "@nivo/radar": "^0.99.0",
+    "@tanstack/react-query": "^5.81.5",
+    "@tanstack/react-query-devtools": "^5.81.5",
     "axios": "^1.10.0",
     "lucide-react": "^0.513.0",
     "next": "15.3.3",
